+++ commander/github-project.cmd/list.action.ts
@@ -1,0 +1,36 @@
+import listCmd from "./list.cmd";
+import util from "../util";
+import type { ProgramGlobalOptions } from "../index.options";
+import type { ListCmdOptions } from "./list.options";
+import listOptions from "./list.options";
+import logger from "../../helpers/logger";
+import GithubClient from "../../services/github";
+
+const logName = "github-project.list";
+
+export default async function(): Promise<void> {
+	const globalOpts = listCmd.optsWithGlobals<ProgramGlobalOptions>();
+	const opts = listCmd.opts<ListCmdOptions>();
+	util.checkOneOfOptions(opts, listOptions.organization, listOptions.user);
+	if (globalOpts.verbose) logger.enableVerboseMode();
+	const ghToken = util.getGithubToken(globalOpts);
+
+	if ("user" in opts) {
+		util.error("not implemented yet :(");
+	} else {
+		logger.debug(logName, "parsed args", { opts });
+
+		logger.debug(logName, "creating github client");
+		const githubClient = new GithubClient(ghToken);
+
+
+		logger.debug(logName, "retrieving organization projects");
+		const [project, err] = await githubClient.projects.listOrganizationProjects(opts.org);
+
+		if (err) throw err;
+
+		if (!logger.verbose()) console.log(JSON.stringify(project))
+		else logger.debug(logName, "success", project);
+		return util.success();
+	}
+}